+++ lib/webview/src/component/ChatInput.tsx
@@ -6,7 +6,8 @@ export const ChatInput: React.FC<{
   text: string;
   onChange?: (text: string) => void;
   onEnter?: (text: string) => void;
-}> = ({ placeholder, disabled, text, onChange, onEnter }) => {
+  onShiftEnter?: (text: string) => void;
+}> = ({ placeholder, disabled, text, onChange, onEnter, onShiftEnter }) => {
   // callback to automatically focus the input box
   const callbackRef = useCallback((inputElement: HTMLTextAreaElement) => {
     if (inputElement) {
@@ -37,10 +38,11 @@ export const ChatInput: React.FC<{
         }}
         // capture onKeyDown to prevent the user from adding enter to the input
         onKeyDown={(event) => {
-          // ignore shift+enter to allow the user to enter multiple lines
+          // enter sends enter, shift+enter creates a new line
           if (
+            onEnter != null &&
             event.key === "Enter" &&
-            !event.shiftKey &&
+            !(event.shiftKey || event.ctrlKey) &&
             event.target instanceof HTMLTextAreaElement
           ) {
             const value = event.target.value.trim();
@@ -49,6 +51,22 @@ export const ChatInput: React.FC<{
               event.preventDefault();
               event.stopPropagation();
               onEnter?.(value);
+            }
+          }
+
+          // shift-enter sends enter, enter creates a new line
+          if (
+            onShiftEnter != null &&
+            event.key === "Enter" &&
+            (event.shiftKey || event.ctrlKey) &&
+            event.target instanceof HTMLTextAreaElement
+          ) {
+            const value = event.target.value.trim();
+
+            if (value !== "") {
+              event.preventDefault();
+              event.stopPropagation();
+              onShiftEnter?.(value);
             }
           }
         }}